fix(scripts): clarified unit test filenames
